Move joinProgramInvite to PublicPorgramController
